+++ ivy/functional/frontends/tensorflow/general_functions.py
@@ -1,5 +1,7 @@
 # local
 import ivy
+from .. import versions
+from ivy.func_wrapper import with_unsupported_dtypes
 
 
 def clip_by_value(t, clip_value_min, clip_value_max):
@@ -12,3 +14,10 @@ def clip_by_value(t, clip_value_min, clip_value_max):
     )
     t = ivy.array(t)
     return ivy.clip(t, clip_value_min, clip_value_max)
+
+
+@with_unsupported_dtypes(
+    {"2.9.0 and below": ("float16", "bfloat16")}, versions["tensorflow"]
+)
+def eye(num_rows, num_columns=None, batch_shape=None, dtype=ivy.float32, name=None):
+    return ivy.eye(num_rows, num_columns, batch_shape=batch_shape, dtype=dtype)
