+++ ivy_tests/test_ivy/test_frontends/test_tensorflow/test_activations.py
@@ -131,7 +131,7 @@ def test_tensorflow_gelu(
         frontend="tensorflow",
         fn_tree="keras.activations.gelu",
         x=x[0],
-        approximate=approximate,
+        approximate=False,
     )
